requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
frozenlist==1.3.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
python-dotenv==0.20.0
slack-sdk==3.17.0
Werkzeug==2.1.2
yarl==1.7.2
